requirements.txt
absl_py==1.3.0
bart_py==2.0.0
cfl==1.3.1
cupy==13.3.0
h5py==3.7.0
icecream==2.1.3
ml_collections==0.1.1
numpy==1.23.4
opencv_python==4.6.0.66
scikit_image==0.19.3
scipy==1.14.1
sigpy==0.1.23
tensorflow_datasets==4.7.0
tf_nightly_gpu==2.12.0.dev20221119
torch==1.13.0
torchmetrics==0.11.0
tqdm==4.64.1
